+++ app/src/androidTest/java/com/training/trainingapp/view/SignUpContentTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.training.app.trainingapp.utils.TestTags
-import com.training.app.trainingapp.main.view.register.SignUpContent
+import com.training.app.trainingapp.main.view.signup.SignUpContent
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
