+++ app/src/main/java/com/training/app/trainingapp/main/view/navigation/MainApp.kt
@@ -7,7 +7,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.training.app.trainingapp.main.view.forget_password.ForgetPasswordScreen
-import com.training.app.trainingapp.main.view.register.SignUpScreen
+import com.training.app.trainingapp.main.view.signup.SignUpScreen
 import com.training.app.trainingapp.main.view_model.ForgetPasswordViewModel
 import com.training.app.trainingapp.main.view_model.SignUpViewModel
 import com.training.app.trainingapp.utils.Screen
