+++ app/src/main/java/com/training/app/trainingapp/main/view/signup/SignUpContent.kt
@@ -1,4 +1,4 @@
-package com.training.app.trainingapp.main.view.register
+package com.training.app.trainingapp.main.view.signup
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
